--- Admin.Core.Service/Admin/Organization/Output/OrganizationGetOutput.cs
@@ -1,8 +1,0 @@
-﻿using Admin.Core.Service.Admin.Organization.Input;
-
-namespace Admin.Core.Service.Admin.Organization.Output
-{
-    public class OrganizationGetOutput : OrganizationUpdateInput
-    {
-    }
-}
