+++ Admin.Core.Service/Personnel/Organization/Output/OrganizationGetOutput.cs
@@ -1,0 +1,8 @@
+﻿using Admin.Core.Service.Personnel.Organization.Input;
+
+namespace Admin.Core.Service.Personnel.Organization.Output
+{
+    public class OrganizationGetOutput : OrganizationUpdateInput
+    {
+    }
+}
