+++ Admin.Core/Controllers/Personnel/AreaController.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace Admin.Core.Controllers.Personnel
+{
+    /// <summary>
+    /// 域控制器
+    /// </summary>
+    [Area("Personnel")]
+    public abstract class AreaController : BaseController
+    {
+    }
+}
